+++ lib/glimmer/swt/widget_proxy.rb
@@ -26,7 +26,7 @@ require 'glimmer/swt/swt_proxy'
 require 'glimmer/swt/display_proxy'
 require 'glimmer/swt/dnd_proxy'
 require 'glimmer/swt/image_proxy'
-require 'glimmer/swt/java_properties'
+require 'glimmer/swt/properties'
 require 'glimmer/swt/custom/drawable'
 
 # TODO refactor to make file smaller and extract sub-widget-proxies out of this
@@ -44,7 +44,7 @@ module Glimmer
     # Follows the Proxy Design Pattern
     class WidgetProxy
       include Packages
-      include JavaProperties
+      include Properties
       include Custom::Drawable
 
       DEFAULT_STYLES = {